Imported functions from utils folder to complete application in index.js
