+++ .env
@@ -1,3 +1,3 @@
-DB_NAME='employees_db'
-DB_USER='root'
-DB_PASSWORD='happycoding'
+# DB_NAME='employees_db'
+# DB_USER='root'
+# DB_PASSWORD='happycoding'
